Condense classes clean up methods create Helpable module
